# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UKM Politeknik",
        "description": "API UKM Politeknik Negeri Lampung. Response dapat berubah sewaktu waktu menyesuaikan perbaikan",
        "contact": {
            "email": "leopradana013@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/private/users/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Profile",
                "operationId": "620609d51ff16742c343a7aae8de1581",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Perbarui akun",
                "description": "Perbarui informasi akun",
                "operationId": "2c846052ea01d06a74aec2e5d0052134",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "position"
                                ],
                                "properties": {
                                    "foto_profile": {
                                        "description": "Foto Profile",
                                        "type": "file"
                                    },
                                    "name": {
                                        "description": "Nama",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "position": {
                                        "description": "Jabatan Enum: (ketua, wakil, anggota)",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/private/users/tambah-anggota": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Membuat Anggota UKM",
                "description": "Membuat anggota UKM baru",
                "operationId": "3e6d5e24d8048e010ed28afe51cd3535",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id_ukm",
                                    "foto_profile",
                                    "name",
                                    "email",
                                    "position",
                                    "password"
                                ],
                                "properties": {
                                    "foto_profile": {
                                        "description": "Foto Profile",
                                        "type": "file"
                                    },
                                    "id_ukm": {
                                        "description": "Id UKM",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Nama",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "position": {
                                        "description": "Jabatan Enum: (ketua, wakil, anggota)",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/private/users/delete-anggota/{id_users}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Menghapus anggota UKM",
                "description": "Menghapus informasi anggota UKM",
                "operationId": "b17895b6feaf90a0281c1345b249fabd",
                "parameters": [
                    {
                        "name": "id_users",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "_id": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}